fix: SankeyLink.value float + garbage quote-line filter for JSON repair
1. SankeyLink.value: int -> float — ransom-flow links_by_amount puts
   monetary sums (floats) as the value; Pydantic rejected them with 2
   validation errors causing 500 on /api/analytics/ransom-flow.

2. Enrichment JSON repair: extend garbage-line filter to also drop lines
   that are just an opening quote followed by only non-ASCII characters
   (e.g. '    "极速赛车开奖结果历史'). These pass the previous structural-
   char check (they contain '"') but after nuclear non-ASCII strip become
   '    "' — an unclosed string that still breaks json.loads.

+++ src/edu_cti/api/models.py
@@ -621,7 +621,7 @@ class SankeyNode(BaseModel):
 class SankeyLink(BaseModel):
     source: str
     target: str
-    value: int
+    value: float
 
 class AttackFlowResponse(BaseModel):
     """Sankey flow: Attack Vector → Category → Impact Outcome."""
